requirements.txt
requests
python-dotenv
inquirer
flask